subtrees/jetson/requirements.txt
ajaxuploader==0.3.8
aldryn-boilerplates==0.7.7
aldryn-common==1.0.4
aldryn-search==0.3.0
amqp==2.3.2
amqplib==1.0.2
anyjson==0.3.3
Babel==2.5.3
BabelDjango==0.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==2.1.3
blessings==1.7
bpython==0.17.1
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
chronograph==1.0.0
contextlib2==0.5.5
curtsies==0.3.0
daemon==1.1
Django==1.8.18
django-appconf==1.0.2
django-autocomplete-light==2.3.3
django-bootstrap-pagination==1.6.4
django-celery==3.0.9
django-classy-tags==0.8.0
django-cms==3.4.5
django-compress==1.0.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.8
django-formtools==2.1
django-grappelli==2.6.5
django-haystack==2.6.1
django-js-asset==1.1.0
django-mobile==0.7.0
django-mptt==0.9.0
django-picklefield==1.0.0
django-pipeline==1.6.14
django-reversion==1.7
django-rosetta==0.7.14
django-sekizai==0.10.0
django-snippetscream==0.0.7
django-sortedm2m==1.5.0
django-spurl==0.6.5
django-standard-form==1.1.1
django-tastypie==0.14.0
django-tinymce==2.7.0
django-treebeard==4.2.0
djangocms-admin-style==1.2.8
et-xmlfile==1.0.1
eventbrite==3.3.3
futures==3.2.0
greenlet==0.4.14
guess-language==0.2
html5lib==1.0.1
idna==2.6
initd==0.0.1
jdcal==1.3
kombu==4.1.0
lxml==4.1.1
mailsnake==1.6.4
microsofttranslator==0.8
mimeparse==0.1.3
MySQL-python==1.2.5
newrelic==2.106.0.87
oauthlib==2.0.6
openpyxl==2.5.0
ordereddict==1.1
Pillow==5.0.0
pipdeptree==0.11.0
polib==1.1.0
poster==0.8.1
pyDes==2.0.1
pyExcelerator==0.6.4.1
Pygments==2.2.0
pyoembed==0.1.2
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.3
raven==6.5.0
recaptcha==1.0rc1
requests==2.19.1
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.4
staticgenerator==1.4.1
twython==3.6.0
typing==3.6.6
Unidecode==0.4.21
urllib3==1.22
URLObject==2.4.3
vine==1.1.4
vobject==0.9.5
wcwidth==0.1.7
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
